backend/users/serializers.py
from rest_framework import serializers
from .models import User, Business

class BusinessSerializer(serializers.ModelSerializer):
    class Meta:
        model = Business
        fields = ["id", "name", "address", "phone", "timezone", "created_at"]

class UserSerializer(serializers.ModelSerializer):
    business = BusinessSerializer(read_only=True)
    class Meta:
        model = User
        fields = ["id", "username", "email", "first_name", "last_name", "role", "business"]
